py-capture/requirements.txt
# Empty
pillow
pySerial
